not using 100 percent while idle
Made a change a bit ago to not re-render when nothing had changed...
but that made the mainloop idle at 100% cpu!
Becase the swap interval was zero (used for games) and it polled events.
Now interval is one, so it will not run faster than vsync,
and it waits for events before even trying to render (with a timeout).
So, idle is close to zero.  Only when processing data does it use cpu.